--- src/main/java/com/example/whiplash/article/domain/document/Category.java
@@ -1,13 +1,0 @@
-package com.example.whiplash.article.domain.document;
-
-import lombok.Getter;
-
-@Getter
-public enum Category {
-    BATTERY,
-    MEDICINE,
-    STEEL,
-    GOLD,
-    GENERAL,
-    ;
-}
