--- src/main/java/com/example/whiplash/article/domain/entity/SummarizedArticleIndex.java
@@ -1,22 +1,0 @@
-package com.example.whiplash.article.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-@Table(name = "article_indices")
-@Entity
-public class SummarizedArticleIndex {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String title;
-    private String press;
-    private LocalDateTime publishedAt;
-    private LocalDateTime createdAt;
-    private String originalMongoId;
-}
