+++ src/main/java/com/example/whiplash/article/original/web/dto/response/ArticleListItemResponse.java
@@ -1,0 +1,9 @@
+package com.example.whiplash.article.original.web.dto.response;
+
+import java.time.LocalDateTime;
+
+public record ArticleListItemResponse(
+    String id,
+    String title,
+    LocalDateTime publishedAt
+) {}
