--- src/main/java/com/example/whiplash/article/service/ArticleTaskProducer.java
@@ -1,7 +1,0 @@
-package com.example.whiplash.article.service;
-
-import java.time.LocalDateTime;
-
-public interface ArticleTaskProducer {
-    String produce(String articleId, LocalDateTime timestamp);
-}
