+++ src/main/java/com/example/whiplash/article/summary/domain/document/Category.java
@@ -1,0 +1,13 @@
+package com.example.whiplash.article.summary.domain.document;
+
+import lombok.Getter;
+
+@Getter
+public enum Category {
+    BATTERY,
+    MEDICINE,
+    STEEL,
+    GOLD,
+    GENERAL,
+    ;
+}
