+++ src/main/java/com/example/whiplash/article/summary/service/ArticleTaskProducer.java
@@ -1,0 +1,7 @@
+package com.example.whiplash.article.summary.service;
+
+import java.time.LocalDateTime;
+
+public interface ArticleTaskProducer {
+    String produce(String articleId, LocalDateTime timestamp);
+}
