--- src/main/java/com/example/whiplash/article/web/dto/response/ArticleDetailResponse.java
@@ -1,7 +1,0 @@
-package com.example.whiplash.article.web.dto.response;
-
-import java.util.List;
-
-public record ArticleDetailResponse(
-    List<ArticleSummaryDTO> summaries
-) {}
