--- src/main/java/com/example/whiplash/article/web/dto/response/ArticleListItemResponse.java
@@ -1,9 +1,0 @@
-package com.example.whiplash.article.web.dto.response;
-
-import java.time.LocalDateTime;
-
-public record ArticleListItemResponse(
-    String id,
-    String title,
-    LocalDateTime publishedAt
-) {}
