--- src/main/java/com/example/whiplash/article/web/dto/response/ArticleSummarizationResponse.java
@@ -1,33 +1,0 @@
-package com.example.whiplash.article.web.dto.response;
-
-import com.example.whiplash.article.service.ArticleRegisterInfo;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ArticleSummarizationResponse {
-
-    private boolean success;
-    private int processedCount;
-    private List<String> failedIds;
-    private List<String> jobIds;
-    private String message;
-
-    public static ArticleSummarizationResponse create(boolean success, ArticleRegisterInfo registerInfo, String message) {
-        return ArticleSummarizationResponse.builder()
-                .success(success)
-                .processedCount(registerInfo.processedArticleIds().size())
-                .failedIds(registerInfo.failedIds())
-                .jobIds(registerInfo.jobIds())
-                .message(message)
-                .build();
-    }
-
-}
