--- src/main/java/com/example/whiplash/article/web/dto/response/ArticleSummaryDTO.java
@@ -1,14 +1,0 @@
-package com.example.whiplash.article.web.dto.response;
-
-import com.example.whiplash.article.domain.document.Category;
-import com.example.whiplash.domain.entity.history.email.SummaryLevel;
-
-import java.time.LocalDateTime;
-
-public record ArticleSummaryDTO(
-    String title,
-    Category category,
-    String summarizedContent,
-    SummaryLevel summaryLevel,
-    LocalDateTime publishedAt
-) {}
