--- src/main/java/com/example/whiplash/article/web/dto/response/SummarizationJobStatusDTO.java
@@ -1,23 +1,0 @@
-package com.example.whiplash.article.web.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SummarizationJobStatusDTO {
-
-    private String jobId;
-    private String articleId;
-    private String status;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private int attempts;
-    private String errorMessage;
-}
